+++ Packages/Netherlands3D/VISSIM/Runtime/Data.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Netherlands3D.VISSIM
+{
+    /// <summary>
+    /// Class containing VISSIM data
+    /// </summary>
+    public class Data
+    {
+        
+    }
+}