+++ Packages/Netherlands3D/VISSIM/CHANGELOG.md
@@ -5,4 +5,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [1.0.0] - 2022-02-14
 ### Added
+- VISSIM file setup
 - Old VISSIM code